requirements.txt
torch>=2.0
numpy>=1.24
pandas>=2.0
openpyxl>=3.1
pyyaml>=6.0
matplotlib>=3.7
tqdm>=4.65
